feat(azure): add support for merge-rebase
Azure DevOps has support for semi-linear rebase (rebase with merge
commit, see [1]). Our branch policies enfore this option, which is
incompatible with any of the automergeStrategy currently available.

This adds the `merge-rebase` option, which is different from using
`rebase` (again, see [1]).

When this option is set to auto, Renovate Bot will actually select
semi-linear rebasing if that is the only option. But if there are
multiple (allowed) options, then there is nothing to force it.

Based on closer inspection, I see that Gitea will use merge rebase
when the option is rebase. I did not adapt this, because that would
break current implementation. Also, I only have the option to test
this with Azure DevOps.

[1]: https://devblogs.microsoft.com/devops/pull-requests-with-rebase/